Some fixes (#80)
* fix(config): accommodate optional device_id parameter for mqtt

That parameter was introduced in d7d4244ae2ef2574612f5ff1752020efafabe1b6
to simplify enabling of non-TLS MQTT connections. However, it broke
the `config` command of the CLI, since the ConfigParser class does not
allow null entries by default, and device_id is an optional parameter.

* refactor(config): use the type annotations for parsing and validation

This commit replaces the ad-hoc, field-specific parsing code, with
usage of the type hints already specified in schemas.py. This makes
the parsing code generic, and enables using Pydantic's validation
behavior when parsing the configuration file.

* docs(config): clarify usage doc of config command

* feat(config): add 'unset' command

This enables the removal of a nullable parameter, such as 'device_id'
in the 'mqtt' section. It errors when attempting to remove the value
of a non-nullable parameter (as of now, every other parameter besides
'device_id').