chore(readme): consolidate prerequisites, prepare version matrix
This rearrangement of subsections (mostly in the prerrequisites
section) intends to keep procedures for each prerrequisite bound
to a single subsection. Notably, the subsections for the agent
and mosquitto have been simplified.

The instructions for wasi-sdk have been removed in this commit,
just to be reworked in the next one.