chore(schemas): constrain type for ip address
This commit leverages Pydantic's field validation behavior
for ensuring that the program will always refuse specified
IP addresses or host names that don't conform to the IP protocol.

Relying on this validation behavior makes it necessary to
update the unit tests, so that Hypothesis does them without
issuing invalid addresses (otherwise, the tests invariably
fail). In this case, the updates required changing the generic
`text` Hypothesis strategy used for injecting strings into the
various parameters, with a more specific string generation
strategy.

Hence, a more verbose, yet simple strategy for generating strings
is added here, which enables constraining the character set for
the first character, and also for the remaining characters.