chore(ci): use default location of https ca certs as cache
When the CLI is launched with a non-default configuration dir,
it looks at the default configuration dir to see if it already
has the trusted HTTPS CA certs bundle there, in which case it
will copy it into the non-default dir. This makes stops
launches of the E2E test from downloading the certs bundle over
and over again.