chore(io): switch mqtt driver to asynchronous i/o
This commit and its ensuing ones make the switch from using
synchronous I/O for the mqtt client Paho, to use async I/O
on top of the Trio library. Not all changes are contained here,
but the ones used by the other classes in the CLI, such as the
commands. Each command will be updated separately in the next
commits.