chore(config): update config implementation with async i/o
Configuration is among the simplest command features of the
CLI, so it is picked first to migrate to the async framework.
It sets the basic pattern for this migration, which is that
the callback of Typer class that defines the command, is
in charge of invoking the async event loop that runs the
main async task (`configure_task`), which instantiates the
MQTT client `Agent`, and runs the configure action within an
established MQTT session, managed by the `mqtt_scope` async
context manager, without MQTT subscriptions. After the
configuration is sent, there is nothing left to do, hence
`async_done()` is called such that the `mqtt_scope` finishes.

Note also that the `configure()` method of `Agent` is now async,
so it runs within an async scope such as `mqtt_scope()`. This is
also a pattern for other commands following. The `configure()`
method is now a thin wrapper of the `publish()` method, for
increasing modularity.