chore(rpc): update rpc implementation with async i/o
RPC is also a simple feature and is used by others, so it is
picked next to migrate to the async framework. Its testing
is however split into behavior within the async loop, and
outside of it (e.g. at the Typer callback level).

Similarly to the configure feature, the `rpc` method of `Agent`
is now async (so must run within an `mqtt_scope`), and the
Typer callback for it now invokes `trio.run()` with an async
task that sets this all up.