chore(logs): update get logs subcommand to async i/o
This is the first mixed MQTT publish and subscribe feature
that is migrated to async I/O. For simplifying the update
to the unit tests, the publish part (which is the RPC command
that enables module log sending) is kept separate from the
subscribe part. Now, since the subscribe part is driven by the
`_loop_forever` method of `Agent` which already runs an `mqtt_scope`
within `trio.run`, the Typer callback for this subcommand runs
the publish task defined by `request_instance_logs` of `Agent` in
a separate `trio.run` loop.

In compensation for the minor added complexity described above,
migrating to async tasks has spared us the need to use a global
variable for the timeout behavior, and it can be replaced with
Trio's provided timeout machinery.