feat(deploy): perform deploy action as a state machine, part 3/3
The logic flow for parsing and preparing the manifest to deploy
is consolidated into the Typer callback for the deploy command,
with the execution of the deployment FSM in its async scope, and
the webserver running as a context manager, which provides for
a clean exit under any conditions.

Once the goal of the FSM is attained, when the `done()` flag is
set by its `verify_report()` method, success is reported and the
async context is `async_done()`.