feat(tls): add configuration for using TLS
TLS in the CLI is enabled by setting the optional parameters
ca_certificate and ca_key to the paths of PEM files containing
the certificate-key pair of a CA under control of the user
(i.e. a self-signed CA), which is to be used for issuing signed
client certificates, both for the CLI itself and for the agent
being spawned, in order to connect to the configured MQTT broker.

The prerrequisite is that the server certificate being used
by the configured MQTT broker, needs to be signed with the user's
CA as well.