feat(tls): configure the mqtt client with an own certificate
We introduce the `cryptography` python library which provides
interfaces to a rich set of SSL objects. In this case, the
MQTT client instantiated by the CLI is configured to use a
fresh client certificate-key pair that is signed by the CA
configured by the user. The pair is generated on the fly,
and is persisted to a location within the CLI's home folder.