feat(tls): make the agent connect over TLS
The `wedge-cli start` command now uses the TLS machinery,
with some relevant additions, to enable the agent connect
to the MQTT broker over TLS, even if the broker is also
running in the same host (localhost). This particular
scenario is made possible by using cwrap's nss_wrapper [1],
which makes it possible for the agent to resolve localhost
to whatever arbitrary Common Name (CN) is present in the
server's subject of its TLS certificate. The server CN is
inquired on the fly from the server, just before launching
the agent.