feat(broker): add mqtt broker as a bundled binary
The rumqtt project provides a MQTT v3.1.1 broker implementation, which
matches the protocol version supported by the agent in the device. It
abides by the Apache-2.0 license, hence bundling its binary is allowed.
Static release builds for Linux and Windows are bundled.

This commit introduces the no-TLS support, and removes handling of
TLS certificate generation from the E2E test.