fix(deploy): generate deployment manifest with local network's IP
As the CLI now errors out if the specified webserver host is not
localhost, it should nevertheless generate the deployment manifest
for the device using the machine's local network IP, so that a
camera can reach it.

This commit is used as an opportunity to define the local_network
submodule, moving the is_localhost() function out of the tls
submodule, as it was only tangentially related to it.