feat(gui): introduce gui mode with home screen
The new 'gui' command starts a GUI based on the Kivy
framework, plus the KivyMD library for simplified layouting
based on Google's Material Design guidelines. Kivy uses
an event loop that here is managed with Trio, with the
business logic of the GUI being implemented by the Driver
class.

Although not strictly required, we follow the Model-View-Controller
structure for connecting the logic and the view with the business
logic, as a way to provide a clearer separation of concerns.

Given that the Kivy is originally synchronous (with compatibility
with async event loops such as Trio), a sync-async binding is
provided by the SyncAsyncBridge class, in order to trigger events
within the Kivy event loop from events coming from the network.
This will be used in upcoming commits implementing new screens.