feat(gui): add streaming screen
This new screen implements streaming control of the camera,
with a basic text-mode, syntax-highlighted display of the
inference results from the camera.

As a shortcut for demonstration purposes, the directory where
images and inference results are uploaded is a random temporary
one allocated when the GUI starts up. As a usability trade-off,
the path to the directory is copied to the clipboard, so that
the user can open it with the system's file explorer.

There is a partial implementation for selecting a Region of Interest
(ROI) on the image. ROI acquisition is started when clicking a button,
but there is currently no usage for the acquired ROI.
-