refactor(gui): adjust relative proportions and use correct widget
The size_hint* properties are arbitrary units that serve for
setting proportions relative to the extent of the dimension they
apply to. Hence, it's more intuitive to use them like "set this
widget twice as large as this other, while this third widget be
10% the second's size".

The MDScrollView is not necessary at all since there won't be large
deployment statuses to display.