feat(streaming): introduce two ROI types and their mapping
The two types of ROI specification are raw and normalized:
- Raw are in integer units for pixels, and the IMX500 sensor
pixel matrix dimensions are included as the effective set this
ROI type is based one.
- Normalized are in floating point units for covering the [0, 1]
number set, and will be useful for translating between Kivy's
point coordinate system and the IMX500 ROI specification coordinate
system.

These types and the associated functions will be applied in the
next commits.