refactor(roi): introduce widget active subregion
The active subregion of the ImageWithROI widget is the
portion of the widget which is filled by the contained
image, given that an image is displayed within the widget
area while respecting the image's aspect ratio. This is
relevant as the window, and hence the widget, can be resized
arbitrarily, with the subregion changing accordingly. Having
the active subregion is necessary to:
- Maintain a drawn ROI rectangle across window resizes. That is,
  that even though the image is shrunk or enlarged when resizing,
  the ROI's corners remain on top of the same image features as
  when drawn initially.
- Enable drawing the ROI strictly within the image area, capturing
  a proper subregion of the image.
These behaviors will be added in upcoming commits.

The active subregion is captured into unit coordinates, by
leveraging that the image center and the widget center are
constrained to always coincide.