refactor(streaming): add camera status class
The Camera class is to be updated with information sent by
the camera firmware, such that the GUI can react to the
camera status as updates are reported. For a first state
variable, the streaming status is represented into enum
class StreamStatus, which is used as a member of the model
of the streaming screen.

With this scaffolding, the state of  the controls in the
streaming screen are made to react to the stream status as
reported by the camera, so that the user is not led into
inconsistent states. Behavior of the controls are updated
in the next commit.