feat(application): add screen for deploying wasm apps
The screen provides a file selector for the user to pick
an AOT-compiled WASM module for the Xtensa architecture.
The file selector is opened by clicking a button with a
file cog icon.

A simple header validation is performed to enforce that the
selected file is Xtensa AOT WASM module. If successful, the
"deploy" button is enabled (if camera readyness is asserted)
and when triggered, a deployment operation is carried out.

Separately, the deployment status report from the device is
displayed in the text widget that takes up most of the area
of this screen. An additional label widget is used to show
the reconcile status of the deployment operation, giving the
user a live indication of the ongoing deployment operation.