feat(application): make camera readiness depend on periodic updates
The readiness criterion is made more strict, by requiring that
the camera be sending status updates at a fixed time interval
(defaults to three seconds). In order to enforce this, an extra
background task is created, which waits for a timer to expire
and if it does, will send the camera a configuration message that
should make it send reports on the periodic interval.

On every incoming status message, the timer deadline is extended,
such that during expected operation, the timer only expires during
the very first time a camera connects, or during disconnections.

The reason for this change, is because it was determined necessary
for the deployment state machine to proceeed reliably through the
required steps, as it is driven by deployment status updates from
the camera.