feat(agent): add mechanism for determining on-wire schema version
The determine_onwire_schema() method is an async routine for
awaiting for the first available 'systemInfo' report from the
camera, which contains the 'protocolVersion' field that holds the
on-wire schema version that the camera is currently configured for.

The 'onwire_schema' data member of Agent holds the result from the
readout of the schema version. In this commit, only the configure()
method is modified, such that it encodes the payload of the
configuration message as base64 only if the determined schema version
is EVP1, according to the agent source linked in the comment.