fix(mqtt): avoid trio generator error
Prior to this commit, usage of the messages() async generator
caused a "Corrupt Cancellation Stack" exception in the parent
trio nursery. The details are discussed in long-standing issue
https://github.com/python-trio/trio/issues/638. An adequate
trade-off solution is provided by the trio-util library, which
requires a minor code change on every usage of the messages()
generator.

However, seizing the unavoidable change in indentation that the
fix requires, the following commits will apply the fix to each
separate reference of messages(), looking to simplify the Agent
class.