chore(agent): initialize on-wire schema version from CLI config
This commit and the next ones will remove the initialization of
the on-wire schema from the device's systemInfo message, using
instead the value set in the CLI configuration.

This is done out of the necessity to support ancient firmware
versions which do not include the 'protocolVersion' field in the
systemInfo message.