feat(deploy): make deployment not wait for frequent updates for EVP1
EVP1 does not support setting the status report interval period
limits, which makes a deployment operation potentially take very
long if the deployment state machine is not transitioned past its
initial states. Hence, a timeout behavior is defined which will
force the machine state to be the one when it pushes the deployment
manifest to the device, and from then on it proceeds as usual.