feat(camera): implement disconnect timeout behavior
Taking into account that the MQTT protocol is fully asynchronous,
that camera connection status is defined by a simplistic criterion,
that relies on the ability to make the camera emit periodic reports.
If the camera has not emitted reports for over six period intervals,
(20 seconds when the period interval is 3 seconds), then we regard
the camera as disconnected.