fix(broker): ensure process logs are passed to our logging
The broker process log streams were not being flushed after
the first lines from the streams were used to determine that
start up is successful. Not flushing them was causing the
process to lock up after some minutes, which resulted in the
camera being effectively disconnected off the MQTT link.