refactor(gui): use different widgets for file manager
The widget layout that composes the MDFileManager class from
KivyMD exhibits a layouting bug whereby its close button is
not fixed at the top right corner of the modal view, but is
position with a fixed padding right of the title, so when the
selected path is long enough, it goes out of the view, so it
becomes unreachable. This is a bug in the MDTopAppBar class
stack. Hence, we replace this default layouting in our subclass
FileManager