refactor(gui): improve async task loop structure
We simplify the async code in the GUI App class and the
constructor of the Driver class, by making Driver's main()
method the owner of the principal trio Nursery, and hence
the App's app_main() method simply awaits its execution.

Within Driver, we do better structured concurrency by
spawning new trio Nursery instances instead of sharing a
one as a class member, which better conveys the concurrency
structure, and should simplify error handling, with the
end goal of enabling the GUI to handle all errors gracefully
and provide visual indication of any errors for which the
user would benefit receiving feedback.