refactor(gui): enable late initialization of path selector combo
Extra properties and a new event are defined in the PathSelectorCombo
class such that the specification of each combo is fully contained at
the point of instantiation, which in turn, makes the configuration view
free from early initializations of its path selector combo instances.

This is motivated by the rare occurrence of failures during GUI start,
in which the configuration screen was unable to be built, because
assignments to the combos' file manager attributes failed.