refactor(main): build python wheel as a central artifact
Doing `pip install` from the repository clone has proved to be fragile,
as its outcome depends on whether the current user is root, and whether
the installation mode is editable or not, and this has caused issues on
several machines. Hence, a wheel is built such that its contents are
guaranteed to be installed on any target machine.