feat(fstools): use 'watchdog'  to notify of directory deletions
The 'watchdog' library implements file system monitoring
across the major desktop platforms. It is used here to implement
a class that monitors directories in the file system, executing a
given callback when a watched directory is deleted. It will
be leveraged by the Driver class to notify the user when either
of the configured directories is deleted.