refactor(gui): introduce camera state proxy class
This class intends to hold the complete representation
of the camera state into Kivy properties, which are central
to maintain a consistent GUI state at all times. This first
commit migrates the application-level properties into the
class, and upcoming commits will migrate other properties
currently scattered on model classes defined for each of the
views.