refactor(camera): device configuration processing, part 1
This commit starts to migrate the device_config member of several
view models, by adding it first into the CameraState class as a
TrackingVariable instance.

As a tracked variable, the observer subscription pattern is used
to simplify the semantics that implement the factory reset
assurance feature.

Next commit will make trivial the Home screen model, as a first
example of how this model consolidation happens.