feat(camera): add binding method for tracked variables in CameraState
The bind_proxy() method is meant to setup a data binding in the
proxy-property --> state variable direction. The main motivation,
is to enable setting values in members of CameraState, from events
originating from interactions of the user with the GUI.

The first usage of this facility comes in the next commit: adding
the AI Model screen's file location as a tracked state variable,
and creating a corresponding proxy property, so that the state
variable stores the selected AI model file location when the user
picks it on the screen.