refactor(ai_model): part 1/4: remove device_config duplication
This commit removes from the model class for the AI model screen the
properties that performed logic on the device_config member of the
camera state class. It can do so by subscribing to updates of the
proxy property for device configuration during __init__().

Given that the device_config logic is removed from the model class
associated to the AI model screen, the signaling done by the
ota_event objects is moved over to the camera state class. This has
the benefit of moving a part that is not a direct concern of the GUI,
into the camera state class.