refactor(ai_model): part 2/4: set up model file as state variable
The AI model file location is a camera state variable that is an
example of a variable whose value is not generated by the camera
(i.e. it does not come via the MQTT channel), but is given by the
GUI user instead. Hence, this commit introduces:
- The tracking variable in the camera state class
- A proxy property in the GUI's camera proxy class
- The binding of these two, via the bind_proxy() method

Notice that the model_file member of AIModelScreenModel is not
removed yet, because its setter method is the one that computes
the model_file_valid property. This property is treated in the
next commit.