refactor(ai_model): part 3/4: set up model_file_valid as state variable
Validity of the chosen AI model file is made into another state
variable with a bound proxy property. This enables recovering the
file validation chaining, including presenting the error message
pop-up, and computing the disabled status of the "Deploy" button.
I chose to define a boolean property on the view class for one
of the factors in the conditional that computes the disabled state
of the Deploy button, as that allows the computation to be set
in the screen's .kv file, which automatically sets up the event
chaining.

These changes end up turning the model class for the AI model
screen into a trivial subclass of BaseScreenView, as was the case
for the home screen. However, the controller class for the AI model
screen is still referring to members of the now trivial model class.
It is addressed next.