refactor(ai_model): part 4/4: move deployment into core
Given that previous commits moved all variables involved in the
deployment state machine of an AI model, into the CameraState
class, this state machine is no longer bound to objects in the
screen's model or view. Hence, this commit moves it into a
new submodule of /core/camera, and the remaining code in the
controller class is drastically reduced. The unit test now
becomes for the deployment state machine, so its location within
the tests source tree is moved accordingly.