refactor(firmware)[1/8]: remove device_config duplication
This commit removes from the model class for the AI model screen the
properties that performed logic on the device_config member of the
camera state class. It can do so by subscribing to updates of the
proxy property for device configuration during __init__().

This pattern is the same as was applied for the AI model screen view
and model classes.

Also, note that this commit is not correct as it is a line-pick
of a bigger set of changes. It is made here only to illustrate this
specific change related to the device_config variable.

Signed-off-by: Luis Linares <luis@midokura.com>