refactor(firmware)[2/8]: add firmware file variables to camera state
This commit promotes the firmware file properties out of the view
class, and into the camera state class and its proxy properties,
which are bound by a new internal method in `Driver`.

Note that some model properties remain in the model class of the
firmware screen. Those properties are only used to render values
that are computed from the ones included in this commit, which
don't need to belong as separate state variables. Hence, they are
treated differently in next commits.

Also, this commit does not replace the logic around the deleted
model properties. This will be done next.

Signed-off-by: Luis Linares <luis@midokura.com>