refactor(firmware)[8/8]: remove transient model properties
The view class of the firmware screen uses a subclass of TransientStatus
to represent those properties remaining in the screen's model class. The
subclass leverages Kivy's event dispatch system via properties, to enable
the abstract firmware state machine reflecting the ongoing status of the
operation in the screen's widgets for the convenience of the user.

Signed-off-by: Luis Linares <luis@midokura.com>