refactor(test_driver): make mocked driver+agent fixture
It is available both as a simple context manager and as a pytest
fixture. It is necessary to have available as a context manager,
because Hypothesis cannot support pytest fixtures the way a
developer would expect them to work.

This change was necessary because a couple test cases begun
exceeding Hypothesis' timing deadline. To solve that, the common
driver fixture mocks more objects that are immaterial to these test
cases.

Signed-off-by: Luis Linares <luis@midokura.com>