refactor(driver): replace the message task of the FSM
The core idea is that the FSM class only maintains a state, and
its stimulus are provided by an external entity. Hence, the message
task is replaced by the abstract start() method, which enables
implementing the different behavior that EVP1 and EVP2 deployments
exhibit. The MQTT client loop that implements the message task has
been moved to the deploy command implementation submodule, because
for the GUI, stimuli will come from the CameraState as it processes
input from the Driver's mqtt loop.

Signed-off-by: Luis Linares <luis@midokura.com>