refactor(driver): FSM doesn't initialize with an Agent instance
Instead, it gets an async function as an argument that will
be called with the manifest to be deployed. For instantiating, it
will require an OnWireSchema enum instance. This change enables
reusing an existing Agent object in the async context it belongs
to, which makes it easier to use the FSM in different contexts.

The test for the core deploy FSM is updated with the changes done
thus far.

Signed-off-by: Luis Linares <luis@midokura.com>