refactor(deploy): updates to deployment manifest generation
The functions for generating deployment manifest are simplified
leveraging the fact that so far, deployments have been for a
single module instance.

Also, changes to the deployment manifest generation are added
to accommodate that the FSM-owned webserver can have a random
TCP port assigned, while still enabling the deploy CLI command
to use the configured port.

Signed-off-by: Luis Linares <luis@midokura.com>