refactor(gui-deploy): implement deployment based on state variables
This replaces the usage of exec_deployment from the deploy CLI
command implementation, with a set of async state subscriptions
plus a few helper functions, which enables the GUI to reflect a
deployment stage that is consistent with the deployment status
reports emitted by the camera firmware, at all times.

The overall workflow is:

1. The user starts the deployment, which sets deploy_operation
2. That triggers creation of a DeployFSM and calling its start()
3. Subsequent deployment status reports stimulate the FSM, updating
   the state variables (particularly, deploy_stage).
4. Once the FSM finishes (whether by erroring or succeeding), the
   deploy_operation will be set to None, and so will the FSM.

However, even when there is no FSM (i.e. no GUI-initiated deployment
operation), the deploy_stage variable will be updated to reflect the
stage as expected by the GUI user.

Signed-off-by: Luis Linares <luis@midokura.com>