fix: modification to remove bindings from driver
The bindings are removed from driver class and
added as methods to the proxy. As a consequence
the initialization of state and device_manager
were modifed. Now the notification of delete files
is being done from the state which report thorugh a
memory channel to the main thread to trigger the
notification to the user.