feat(devices): make effective name editions in the entries list
The user can type in a name and shortly after they have
stopped typing, the name will be updated everywhere in the
GUI and in persistent configureation. If the user hits the Enter
key, the renaming happens immediately.

A notification gives feedback of the change once it has taken
place.

Signed-off-by: Luis Linares <luis@midokura.com>