refactor(mqtt): handle a broker error gracefully
In order to avoid a broker error from being fatal to Local Console,
the BrokerException raised by spawn_broker must be caught, but since
it is raised from the local nursery, it is wrapped by an ExceptionGroup,
which has somewhat different semantics than a plain exception.

Now, when that happens, the whole MQTT setup should be considered
failed and aborted. However, this async method must be able to send
such information to its caller. This is enabled by the task_status
argument, which is called with an argument that expresses whether the
setup has been successful or not. How this is used by the caller, is
implemented in the next commit.

Finally, upon catching the exception, its text content is sent through
the message channel, for the receiver end of the Driver class to render
as a pop-up notification.

Signed-off-by: Luis Linares <luis@midokura.com>