refactor(config): split add_device() into two steps
Given that device creation now has an intermediate validation for
checking whether the MQTT broker can actually be spawned as per
its configuration, committing to memory such configuration should
only happen after the validation is done.

This splitting will be used for refactoring the device manager on
the next commit.

Signed-off-by: Luis Linares <luis@midokura.com>