refactor(devices-gui): updates to controller for error handling
The Devices Screen controller class is adapted to accommodate the
new async form of the device addition related methods of the
DeviceManager class. The most significant change is that it
uses the Driver's from_sync() method to call the manager's add_device()
async method, and the run_on_ui_thread decorator as the continuation
callback that will render the screen according to the latest
device listing and configuration.

Signed-off-by: Luis Linares <luis@midokura.com>